requirements.txt
discord.py==2.3.1
flask==2.2.5
